[feat] 에러타입 추가
