+++ BaseballGame_weak2/BaseballGame_weak2/Lv2/BaseballExtension.swift
@@ -13,25 +13,8 @@ extension Baseball {
     public mutating func play() {
         print("< 게임을 시작합니다 >")
         
-        // 다운캐스팅으로 self를 복사하여 게임 실행
-        if let game = self as? BaseballGame {
-            // game.currentNumber = createRandomNumber()
-            // Lv3 0을 포함하는 랜덤 수 생성하기
-            game.currentNumber = createRandomNumberLv3()
-            
-            // compareNumber 함수의 결과가 정답인 경우
-            // while 반복을 멈추고 게임 종료
-            while game.rightAnswer != true {
-                print("숫자를 입력하세요")
-                
-                let input = readLine()
-                if let input, let intInput = Int(input) {
-                    game.compareNumber(game.currentNumber!, intInput)
-                } else {
-                    print("올바르지 않은 입력값입니다.")
-                }
-            }
-        }
+        // 게임을 실행하는 메소드
+        tryGame()
     }
 }
